Separate the querySelector data from TDD functions
