+++ src/data.js
@@ -1,0 +1,3 @@
+export let customers = [];
+export let rooms = [];
+export let bookings = [];
