+++ src/scripts.js
@@ -4,23 +4,19 @@
 import './css/styles.scss';
 // An example of how you tell webpack to use an image (also need to link to it in the index.html)
 import './images/turing-logo.png'
-// import './DOMUpdates';
 import { fetchCustomers, fetchRooms, fetchBookings } from './apiCalls';
+import { customers, rooms, bookings } from './data.js' 
+import './DOMUpdates.js'
 //<><><><>Event Listeners
-
-export let customers = [];
-export let rooms = [];
-export let bookings = [];
 
 export function initialize() {
     Promise.all([fetchCustomers(), fetchRooms(), fetchBookings()])
-    .then((value) => {
-        [customers, rooms, bookings] = [...value]
+    .then(([fetchedCustomers, fetchedRooms, fetchedBookings]) => {
+        customers.push(...fetchedCustomers);
+        rooms.push(...fetchedRooms)
+        bookings.push(...fetchedBookings)
       })
       .catch(error => {
-        errorMessage.classList.remove('hidden')
-        errorMessage.innerHTML=`${error}`
         console.error("Error fetching data:", error);
       });
   }
-
